@@ -4,7 +4,7 @@ version: 0.1.7
 repository: https://github.com/ReniDelonzek/select_any
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
